# comanda-api-interna/src/main/openapi/openapi-app-comanda-v1.yml
---
openapi: 3.0.3
info:
  title: comanda-api-interna
  version: 0.1.0
servers:
  - url: ""
tags:
  - name: APP → COMANDA / JMS
    description: "Enviament de missatges a les cues JMS de COMANDA: tasques, avisos\
      \ i permisos. Un missatge és una unitat d'informació publicada a una cua JMS\
      \ perquè sigui processada de manera asíncrona."
paths:
  /api/v1/jms:
    get:
      tags:
        - APP → COMANDA / JMS
      operationId: JmsRestController.index
      description: "Index d'enllaços a les operacions d'enviament de missatges JMS\
        \ disponibles (tasques, avisos i permisos)."
      summary: Llista d'opcions disponibles
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items:
                      $ref: '#/components/schemas/Link'
                  content:
                    type: array
                    items: {}
                  fallbackType:
                    $ref: '#/components/schemas/ResolvableType'
                  fullType:
                    $ref: '#/components/schemas/ResolvableType'
      x-operation-name: index
  /api/v1/jms/avisos:
    post:
      tags:
        - APP → COMANDA / JMS
      operationId: JmsRestController.sendToAvisosQueue
      description: Afegeix un missatge d'avís a la cua JMS 'CUA_AVISOS'.
      summary: Publica un avís a la cua d'avisos
      requestBody:
        content:
          application/json:
            schema:
              description: Representa un avís o notificació publicada a COMANDA
              $ref: '#/components/schemas/Avis'
      responses:
        200:
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
      x-operation-name: sendToAvisosQueue
  /api/v1/jms/permisos:
    post:
      tags:
        - APP → COMANDA / JMS
      operationId: JmsRestController.sendToPermisosQueue
      description: Afegeix un missatge de permisos a la cua JMS 'CUA_PERMISOS'.
      summary: Publica una sol·licitud de permisos a la cua de permisos
      requestBody:
        content:
          application/json:
            schema:
              description: Sol·licitud de permisos per a un usuari o grup sobre un
                objecte
              $ref: '#/components/schemas/Permis'
      responses:
        200:
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
      x-operation-name: sendToPermisosQueue
  /api/v1/jms/tasques:
    post:
      tags:
        - APP → COMANDA / JMS
      operationId: JmsRestController.sendToTasquesQueue
      description: Afegeix un missatge de tipus Tasca a la cua JMS 'CUA_TASQUES'.
      summary: Publica una tasca a la cua de tasques
      requestBody:
        content:
          application/json:
            schema:
              description: Representa una tasca publicada a COMANDA perquè sigui processada
                asíncronament
              $ref: '#/components/schemas/Tasca'
      responses:
        200:
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
      x-operation-name: sendToTasquesQueue
components:
  schemas:
    Avis:
      description: Representa un avís o notificació publicada a COMANDA
      type: object
      properties:
        appCodi:
          description: Codi de l'aplicació que publica l'avís
          type: string
          example: PFI
        entornCodi:
          description: Codi de l'entorn de l'aplicació
          type: string
          example: DEV
        identificador:
          description: Identificador únic de l'avís
          type: string
          example: AV-2025-0001
        tipus:
          description: Tipus d'avís
          $ref: '#/components/schemas/AvisTipus'
        nom:
          description: Títol de l'avís
          type: string
          example: Interrupció programada
        descripcio:
          description: Descripció de l'avís
          type: string
          example: Aturada de manteniment el diumenge a les 8:00
        dataInici:
          description: Data d'inici de vigència
          type: string
          format: date-time
          example: 2025-11-21T08:00:00Z
        dataFi:
          description: Data de fi de vigència
          type: string
          format: date-time
          example: 2025-11-21T10:00:00Z
        redireccio:
          description: URL de redirecció associada a l'avís
          $ref: '#/components/schemas/URL'
          example: https://dev.caib.es/app/avis/AV-2025-0001
        responsable:
          description: Usuari responsable
          type: string
          example: usr1234
        grup:
          description: Grup responsable
          type: string
          example: SUPORT
        usuarisAmbPermis:
          description: Usuaris amb permís de visualització
          type: array
          items:
            type: string
          example: "[\"usr1\", \"usr2\"]"
        grupsAmbPermis:
          description: Grups amb permís de visualització
          type: array
          items:
            type: string
          example: "[\"USUARIS\"]"
    AvisTipus:
      description: Tipus d'avís publicat a COMANDA
      type: string
      enum:
        - NOTICIA
        - INFO
        - ALERTA
        - ERROR
        - CRITIC
    InetAddress:
      type: object
      properties:
        holder:
          $ref: '#/components/schemas/InetAddressHolder'
        canonicalHostName:
          type: string
    InetAddressHolder:
      type: object
      properties:
        originalHostName:
          type: string
        hostName:
          type: string
        address:
          type: integer
          format: int32
        family:
          type: integer
          format: int32
    Link:
      type: object
      properties:
        rel:
          $ref: '#/components/schemas/LinkRelation'
        href:
          type: string
        hreflang:
          type: string
        media:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
    LinkRelation:
      type: object
    Objecte:
      description: Objecte funcional sobre el qual s'apliquen permisos
      type: object
      properties:
        tipus:
          description: Tipus d'objecte
          type: string
          example: EXPEDIENT
        nom:
          description: Nom o títol de l'objecte
          type: string
          example: EXP-12345/2025
        identificador:
          description: Identificador únic de l'objecte
          type: string
          example: 12345
    Permis:
      description: Sol·licitud de permisos per a un usuari o grup sobre un objecte
      type: object
      properties:
        appCodi:
          description: Codi de l'aplicació que sol·licita el permís
          type: string
          example: PORTAFIB
        entornCodi:
          description: Codi de l'entorn de l'aplicació
          type: string
          example: DEV
        usuari:
          description: Usuari al qual s'aplica el permís
          $ref: '#/components/schemas/Usuari'
        grup:
          description: Grup al qual s'aplica el permís
          type: string
          example: GESTORS
        permisos:
          description: Llista de permisos sol·licitats
          type: array
          items:
            type: string
          example: "[\"LECTURA\", \"ESCRIPTURA\"]"
        objecte:
          description: Objecte sobre el qual s'atorguen els permisos
          $ref: '#/components/schemas/Objecte'
        objectesHereus:
          description: Objectes hereus als quals també s'apliquen els permisos
          type: array
          items:
            description: Objecte funcional sobre el qual s'apliquen permisos
            $ref: '#/components/schemas/Objecte'
    Prioritat:
      description: Nivell de prioritat d'una tasca
      type: string
      enum:
        - NONE
        - BAIXA
        - NORMAL
        - ALTA
        - MAXIMA
    ResolvableType:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Type'
        typeProvider:
          $ref: '#/components/schemas/TypeProvider'
        variableResolver:
          $ref: '#/components/schemas/VariableResolver'
        componentType:
          $ref: '#/components/schemas/ResolvableType'
        hash:
          type: integer
          format: int32
        resolved: {}
        superType:
          $ref: '#/components/schemas/ResolvableType'
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/ResolvableType'
        generics:
          type: array
          items:
            $ref: '#/components/schemas/ResolvableType'
        unresolvableGenerics:
          type: boolean
    Tasca:
      description: Representa una tasca publicada a COMANDA perquè sigui processada
        asíncronament
      type: object
      properties:
        appCodi:
          description: Codi de l'aplicació que publica la tasca
          type: string
          example: PORTAFIB
        entornCodi:
          description: Codi de l'entorn de l'aplicació
          type: string
          example: DEV
        identificador:
          description: Identificador únic de la tasca en l'àmbit de l'APP
          type: string
          example: TAS-2025-0001
        tipus:
          description: Tipus funcional de la tasca
          type: string
          example: GENERAR_INFORME
        nom:
          description: Nom curt de la tasca
          type: string
          example: Generar informe mensual
        descripcio:
          description: Descripció detallada de la tasca
          type: string
          example: Generar l'informe mensual de consums per unitat
        estat:
          description: Estat de processament de la tasca
          $ref: '#/components/schemas/TascaEstat'
        estatDescripcio:
          description: Descripció de l'estat actual
          type: string
          example: En cua
        numeroExpedient:
          description: Número d'expedient relacionat (si aplica)
          type: string
          example: EXP-12345/2025
        prioritat:
          description: Prioritat de la tasca
          $ref: '#/components/schemas/Prioritat'
        dataInici:
          description: Data d'inici prevista o real
          type: string
          format: date-time
          example: 2025-11-20T09:30:00Z
        dataFi:
          description: Data de finalització
          type: string
          format: date-time
          example: 2025-11-20T10:00:00Z
        dataCaducitat:
          description: Data de caducitat límit
          type: string
          format: date-time
          example: 2025-12-01T00:00:00Z
        redireccio:
          description: URL de redirecció per accedir a la tasca
          $ref: '#/components/schemas/URL'
          example: https://dev.caib.es/app/tasques/TAS-2025-0001
        responsable:
          description: Usuari responsable
          type: string
          example: usr1234
        grup:
          description: Grup responsable
          type: string
          example: GESTORS
        usuarisAmbPermis:
          description: Llista d'usuaris amb permís
          type: array
          items:
            type: string
          example: "[\"usr1\", \"usr2\"]"
        grupsAmbPermis:
          description: Llista de grups amb permís
          type: array
          items:
            type: string
          example: "[\"GESTORS\", \"SUPORT\"]"
    TascaEstat:
      description: Estat de processament d'una tasca
      type: string
      enum:
        - PENDENT
        - INICIADA
        - FINALITZADA
        - CANCELADA
        - ERROR
    Type:
      type: object
      properties:
        typeName:
          type: string
    TypeProvider:
      type: object
      properties:
        source: {}
        type:
          $ref: '#/components/schemas/Type'
    URL:
      type: object
      properties:
        protocol:
          type: string
        host:
          type: string
        port:
          type: integer
          format: int32
        file:
          type: string
        query:
          type: string
        authority:
          type: string
        path:
          type: string
        userInfo:
          type: string
        ref:
          type: string
        hostAddress:
          $ref: '#/components/schemas/InetAddress'
        handler:
          $ref: '#/components/schemas/URLStreamHandler'
        hashCode:
          type: integer
          format: int32
        tempState:
          $ref: '#/components/schemas/UrlDeserializedState'
    URLStreamHandler:
      type: object
    UrlDeserializedState:
      type: object
      properties:
        protocol:
          type: string
        host:
          type: string
        port:
          type: integer
          format: int32
        authority:
          type: string
        file:
          type: string
        ref:
          type: string
        hashCode:
          type: integer
          format: int32
    Usuari:
      description: Dades bàsiques d'un usuari
      type: object
      properties:
        codi:
          description: Codi únic de l'usuari
          type: string
          example: usr1234
        nom:
          description: Nom complet de l'usuari
          type: string
          example: Anna Serra
        nif:
          description: NIF/NIE de l'usuari
          type: string
          example: 12345678Z
        email:
          description: Correu electrònic
          type: string
          example: anna.serra@example.org
    VariableResolver:
      type: object
      properties:
        source: {}